# Optimize build time