--- a/assets/ext/post.css
+++ b/assets/ext/post.css
@@ -102,7 +102,3 @@
 .post-content blockquote p {
     margin: 0.4em 0;
 }
-
-.post-content abbr {
-    border-bottom: 1px dotted #333;
-}