--- a/lib/metalsmith/image-variation.js
+++ b/lib/metalsmith/image-variation.js
@@ -97,11 +97,8 @@ module.exports = function (options) {
             }
             $ = cheerio.load(files[filePath].contents);
 
-            $('img').each(function () {
-                handleImageAttribute($(this), 'src', files, localRegexp, siteUrl, tasks);
-            });
-            $("meta[name='twitter:image'], meta[property='og:image']").each(function () {
-                handleImageAttribute($(this), 'content', files, localRegexp, siteUrl, tasks);
+            $("img, meta[name='twitter:image'], meta[property='og:image']").each(function () {
+                handleImageAttribute($(this), this.tagName === 'img' ? 'src' : 'content', files, localRegexp, siteUrl, tasks);
             });
         });