--- a/lib/metalsmith/image-variation.js
+++ b/lib/metalsmith/image-variation.js
@@ -89,10 +89,14 @@ module.exports = function (options) {
 
         Object.keys(files).forEach(function (filePath) {
             var $;
-            if ( !filePath.endsWith('.html') ) {
+
+            if ( !filePath.endsWith('.html') || filePath.startsWith('tag') ) {
+                // ignore .xml and tag pages because they contain the same
+                // variations as others pages (post page and post list)
                 return;
             }
             $ = cheerio.load(files[filePath].contents);
+
             $('img').each(function () {
                 handleImageAttribute($(this), 'src', files, localRegexp, siteUrl, tasks);
             });