--- a/lib/metalsmith/image-variation.js
+++ b/lib/metalsmith/image-variation.js
@@ -6,22 +6,13 @@ var cheerio = require('cheerio'),
     imageminOptipng = require('imagemin-optipng'),
     imageminGifsicle = require('imagemin-gifsicle'),
     imageminSvgo = require('imagemin-svgo'),
-    imagePath = '/images/';
-
-function getLocalRegexp(siteUrl) {
-    return new RegExp('(' + imagePath + '|' + siteUrl + imagePath + ')');
-}
-
-function isLocalImage(src, siteUrl) {
-    var localRegexp = getLocalRegexp(siteUrl);
-
-    return src.match(localRegexp);
-}
+    imagePath = '/images/',
+    multipleSlash = /\/+/g;
 
 function normalizeImageSrc(src, siteUrl) {
-    var localRegexp = getLocalRegexp(siteUrl);
-
-    return src.replace(localRegexp, 'images/').replace(/\/+/g, '/');
+    return src.replace(siteUrl, '')
+        .replace(imagePath, 'images/')
+        .replace(multipleSlash, '/');
 }
 
 function getImageVariation(src) {
@@ -34,7 +25,7 @@ function getImageVariation(src) {
 }
 
 function getOriginalImageFile(files, src, variation) {
-    var origSrc = src.replace('/' + variation, '');
+    var origSrc = src.replace(`/${variation}`, '');
 
     return files[origSrc];
 }
@@ -69,11 +60,11 @@ function generateVariation(variation, src, files, done) {
         });
 }
 
-function handleImageAttribute(element, attr, files, siteUrl, tasks) {
+function handleImageAttribute(element, attr, files, localRegexp, siteUrl, tasks) {
     var src = element.attr(attr),
         variation;
 
-    if ( !isLocalImage(src, siteUrl) ) {
+    if ( !src.match(localRegexp) ) {
         return;
     }
     src = normalizeImageSrc(src, siteUrl);
@@ -90,7 +81,8 @@ function handleImageAttribute(element, attr, files, siteUrl, tasks) {
 }
 
 module.exports = function (options) {
-    var siteUrl = options.siteUrl;
+    var siteUrl = options.siteUrl,
+        localRegexp = new RegExp(`(${imagePath}|${siteUrl}${imagePath})`);
 
     return function (files, metalsmith, done) {
         var tasks = [];
@@ -102,10 +94,10 @@ module.exports = function (options) {
             }
             $ = cheerio.load(files[filePath].contents);
             $('img').each(function () {
-                handleImageAttribute($(this), 'src', files, siteUrl, tasks);
+                handleImageAttribute($(this), 'src', files, localRegexp, siteUrl, tasks);
             });
             $("meta[name='twitter:image'], meta[property='og:image']").each(function () {
-                handleImageAttribute($(this), 'content', files, siteUrl, tasks);
+                handleImageAttribute($(this), 'content', files, localRegexp, siteUrl, tasks);
             });
         });