--- a/templates/layout.html
+++ b/templates/layout.html
@@ -17,9 +17,6 @@
     <link rel="apple-touch-icon" sizes="72x72" href="/img/touch-icon-ipad.png">
     <link rel="apple-touch-icon" sizes="114x114" href="/img/touch-icon-iphone-retina.png">
     <link rel="apple-touch-icon" sizes="144x144" href="/img/touch-icon-ipad-retina.png">
-    <!--[if lt IE 9]>
-    <script>(function(){var tags=['header','article','footer','nav', 'time', 'section', 'figure', 'abbr'];for(var i=0;i!=tags.length;i++) document.createElement(tags[i]);})();</script>
-    <![endif]-->
     {% block customMeta %}{% endblock %}
 </head>
 <body>