```
define: 0.171ms
assets: 4.670ms
fileToMetadata: 40.507ms
myth: 163.763ms
ignore: 22.840ms
moment: 35.254ms
tags: 639.684ms
collections: 153.193ms
pagination: 4.514ms
fileMetadata: 19.578ms
metallic: 2523.777ms
markdown: 191.348ms
collectPhotos: 164.032ms
permalinks: 774.173ms
feed: 60.080ms
tagLangFeed: 280.682ms
styleRenamePlugin: 0.086ms
layouts: 3358.075ms
imageVariation: 4386.691ms
htmlMinifier: 7881.534ms
pdfize: 1696.972ms
renamer: 5.741ms
```