@dpobel 2017-11-14
```
define: 0.164ms
assets: 4.106ms
fileToMetadata: 38.941ms
myth: 165.142ms
ignore: 25.245ms
moment: 33.997ms
tags: 651.186ms
collections: 156.284ms
pagination: 4.788ms
fileMetadata: 21.242ms
metallic: 2599.626ms
markdown: 192.889ms
collectPhotos: 181.013ms
permalinks: 781.968ms
feed: 11.167ms
tagLangFeed: 269.028ms
styleRenamePlugin: 0.070ms
layouts: 3640.620ms
* htmlMinifier: 6819.336ms
* imageVariation: 1824.328ms
pdfize: 1404.740ms
renamer: 7.226ms
```